+++ test/integrations/adwords.js
@@ -8,10 +8,10 @@ describe('AdWords', function(){
   var sinon = require('sinon');
 
   var settings = {
-    conversionId: 983265867,
+    conversionId: 978352801,
     events: {
-      signup: '7c8fc3c1',
-      login: '49aa6e21',
+      signup: '-kGkCJ_TsgcQofXB0gM',
+      login: 'QbThCM_zogcQofXB0gM',
       play: 'b91fc77f'
     }
   };
@@ -24,43 +24,80 @@ describe('AdWords', function(){
   it('should have the correct settings', function(){
     test(adwords)
       .name('AdWords')
-      .readyOnInitialize()
+      .readyOnLoad()
       .option('events', {});
   })
 
-  describe('#track', function(){
-    before(function(){
-      sinon.spy(AdWords, 'load');
+  describe('#load', function(){
+    it('should load', function(done){
+      adwords.load(function(){ done(); });
+    })
+  })
+
+  describe('#conversion', function(){
+    beforeEach(function(){
+      var els = document.getElementsByTagName('img');
+      for (var i = 0; i < els.length; ++i) {
+        if (!els[i].src) continue;
+        if (/googleadservices/.test(els[i].src)) continue;
+        els[i].parentNode.removeChild(els[i]);
+      }
     })
 
-    afterEach(function(){
-      AdWords.load.reset();
+    it('should set globals correctly', function(done){
+      adwords.conversion({ conversionId: 1, label: 'baz', value: 9 }, done);
+      assert(1 == window.google_conversion_id);
+      assert('en' == window.google_conversion_language);
+      assert('3' == window.google_conversion_format);
+      assert('ffffff' == window.google_conversion_color);
+      assert('baz' == window.google_conversion_label);
+      assert(9 == window.google_conversion_value);
+      assert(false == window.google_remarketing_only);
+    })
+
+    it('should override document.write', function(done){
+      var write = document.write;
+      adwords.conversion({ conversionId: 1, label: 'baz', value: 9 }, done);
+      assert(write != document.write);
+    });
+
+    it('should restore document.write', function(done){
+      var write = document.write;
+      adwords.conversion({ conversionId: 1, label: 'baz', value: 9 }, function(){
+        assert(write == document.write);
+        done();
+      });
+    })
+  })
+
+  describe('#track', function(){
+    beforeEach(function(done){
+      adwords.on('ready', done);
+      sinon.spy(adwords, 'conversion');
+      adwords.initialize();
     })
 
     it('should not send if event is not defined', function(){
       test(adwords).track('toString', {});
-      assert(!AdWords.load.called);
+      assert(!adwords.conversion.called);
     })
 
-    it('should send if event is found', function(){
-      test(adwords)
-        .track('signup', {})
-        .called(AdWords.load)
-        .with({ value: 0, label: '7c8fc3c1', script: 0}, { id: 983265867 });
+    it('should send event if its defined', function(){
+      test(adwords).track('signup', {});
+      assert(adwords.conversion.calledWith({
+        conversionId: adwords.options.conversionId,
+        label: adwords.options.events.signup,
+        value: 0,
+      }))
     })
 
     it('should send revenue', function(){
-      test(adwords)
-        .track('login', { revenue: '$50' })
-        .called(AdWords.load)
-        .with({ value: 50, label: '49aa6e21', script: 0 }, { id: 983265867 });
-    })
-
-    it('should send correctly', function(){
-      test(adwords).track('play', { revenue: 90 });
-      var img = AdWords.load.returnValues[0];
-      assert(img);
-      assert('http://www.googleadservices.com/pagead/conversion/983265867?value=90&label=b91fc77f&script=0' == img.src)
+      test(adwords).track('login', { revenue: 90 });
+      assert(adwords.conversion.calledWith({
+        conversionId: adwords.options.conversionId,
+        label: adwords.options.events.login,
+        value: 90
+      }));
     })
   })
 })