+++ test/integrations/intercom.js
@@ -30,7 +30,6 @@ describe('Intercom', function () {
       .global('Intercom')
       .option('activator', '#IntercomDefaultWidget')
       .option('appId', '')
-      .option('counter', true)
       .option('inbox', false);
   });
 
@@ -118,7 +117,23 @@ describe('Intercom', function () {
         .with('boot', {
           app_id: settings.appId,
           user_id: 'id',
+          firstName: 'john',
+          lastName: 'doe',
           name: 'john doe',
+          id: 'id'
+        });
+    })
+
+    it('should respect .name, .firstName and .lastName', function(){
+      test(intercom)
+        .identify('id', { firstName: 'john', lastName: 'doe', name: 'baz' })
+        .called(window.Intercom)
+        .with('boot', {
+          app_id: settings.appId,
+          user_id: 'id',
+          firstName: 'john',
+          lastName: 'doe',
+          name: 'baz',
           id: 'id'
         });
     })
@@ -187,8 +202,7 @@ describe('Intercom', function () {
         user_id: 'id',
         id: 'id',
         widget: {
-          activator: '#IntercomDefaultWidget',
-          use_counter: true
+          activator: '#IntercomDefaultWidget'
         }
       }));
     });
@@ -202,8 +216,7 @@ describe('Intercom', function () {
         user_id: 'id',
         id: 'id',
         widget: {
-          activator: '#Intercom',
-          use_counter: true
+          activator: '#Intercom'
         }
       }));
     });